update card after rating
